WIP adding ssr
