+++ package.json
@@ -21,6 +21,8 @@
   "type": "module",
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-register": "^6.26.0",
     "express": "^4.18.2",
     "file-loader": "^6.2.0",
     "formik": "^2.2.9",
@@ -43,6 +45,7 @@
     "@types/react-dom": "^18.0.10",
     "@typescript-eslint/eslint-plugin": "^5.40.1",
     "@typescript-eslint/parser": "^5.40.1",
+    "babel-cli": "^6.26.0",
     "babel-loader": "^8.2.5",
     "cross-env": "^7.0.3",
     "css-loader": "^6.7.1",
