+++ src/server/Component.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Component = () => {
+    return (
+        <div>Component</div>
+    )
+}
+
+export default Component;
